refactor: define "engine/", "core/"
