+++ src/README.md
@@ -15,6 +15,10 @@ Process files are wanted to lint/fix.
 
 - textlint-engine.js
 - rule-manager.js
+- textlint-module-resolver.js
+
+These are shared between config and engine.
+Don't shared between engine and core.
 
 ## Core
 
@@ -23,3 +27,6 @@ Process file/text wanted to lint/fix
 - textlint-core.js
 - source-code.js
 - rule-creator-set.js
+- task/
+- linter/
+- fixer/
