+++ src/core/rule-context.js
