+++ src/core/rule-creator-helper.js
