+++ src/core/rule-error.js
