+++ src/core/source-location.js
