+++ src/fixer/source-code-fixer.js
@@ -1,5 +1,5 @@
 const debug = require("debug")("textlint:source-code-fixer");
-import SourceCode from "../rule/source-code";
+import SourceCode from "../core/source-code";
 const BOM = "\uFEFF";
 
 /**
