+++ test/source-code-fixer-test.js
@@ -2,7 +2,7 @@
 "use strict";
 import assert from "power-assert";
 import SourceCodeFixer from "../src/fixer/source-code-fixer";
-import SourceCode from "../src/rule/source-code";
+import SourceCode from "../src/core/source-code";
 import {parse} from "markdown-to-ast";
 const TEST_CODE = "var answer = 6 * 7;";
 const TEST_AST = parse(TEST_CODE);
